ai_service/requirements.txt
fastapi
uvicorn[standard]
ultralytics
transformers
sentence-transformers
chromadb
openai-whisper
spacy
torch
torchvision
Pillow
python-multipart
pymongo
python-dotenv
numpy
httpx
